requirements.txt
py-cord
aiohttp
pytz
discord-py
py-cord
pytz
flask